refactor(desktop): extract sidebar row subcomponents
AgentsSection ad-hoc/workflow rows and ResolveCluster row moved to

sibling files; export const; 1183 to 922 and 312 to 155 lines.
