+++ apps/desktop/src/features/workspace/components/WorkspacesSidebar/parts/ResolveCluster.tsx
@@ -1,22 +1,8 @@
-import { useState } from 'react';
-import { StatusDot, cn } from '@goodboy/ui';
-import {
-  ArrowUpRight,
-  ChevronDown,
-  ChevronRight,
-  MessageSquareReply,
-  Play,
-  Upload,
-} from 'lucide-react';
+import { ArrowUpRight, ChevronDown, ChevronRight, MessageSquareReply, Play } from 'lucide-react';
 import type { Agent, AgentId, DiffComment, PrComment, SessionId } from '@goodboy/types';
-import { agentHasUnread } from '../../../../../store';
 import { openUrl } from '../../../../../shared/lib/editor';
-import {
-  ResolverStateBadge,
-  resolverBadgeState,
-} from '../../../../session/components/ResolverStateBadge';
-import { CommentSnippet } from '../../../../session/components/CommentSnippet';
 import { resolverStatus, type ResolverState, type ResolverStatus } from '../lib';
+import { ResolveClusterRow } from './ResolveClusterRow';
 
 type ResolveClusterProps = {
   readonly agents: ReadonlyArray<Agent>;
@@ -36,24 +22,7 @@ type ResolveClusterProps = {
   readonly onResolveThread: (threadId: string) => Promise<void> | void;
 };
 
-const diffLocation = (comment: DiffComment): string => {
-  if (comment.anchor == null) {
-    return comment.filePath;
-  }
-  return `${comment.filePath}:${comment.anchor.lineNumber}`;
-};
-
-const commentLocation = (comment: PrComment): string | null => {
-  if (comment.source === 'issue') {
-    return 'conversation';
-  }
-  if (comment.path == null) {
-    return null;
-  }
-  return comment.line != null ? `${comment.path}:${comment.line}` : comment.path;
-};
-
-export function ResolveCluster({
+export const ResolveCluster = ({
   agents,
   sessionId,
   isTaskActive,
@@ -69,7 +38,7 @@ export function ResolveCluster({
   onSelect,
   onForceNext,
   onResolveThread,
-}: ResolveClusterProps) {
+}: ResolveClusterProps) => {
   const statusOf = (a: Agent): ResolverStatus =>
     resolverStatus(a, resolvedThreadIds, pendingThreadIds, resolverState[a.id]);
   const resolvedCount = agents.filter((a) => statusOf(a) === 'resolved').length;
@@ -183,130 +152,4 @@ export function ResolveCluster({
       ) : null}
     </div>
   );
-}
-
-type ResolveClusterRowProps = {
-  readonly agent: Agent;
-  readonly index: number;
-  readonly total: number;
-  readonly status: ResolverStatus;
-  readonly threadComment: PrComment | null;
-  readonly diffComment: DiffComment | null;
-  readonly isSelected: boolean;
-  readonly isTaskActive: boolean;
-  readonly canJump: boolean;
-  readonly onSelect: () => void;
-  readonly onJump: () => void;
-  readonly onResolveThread: (threadId: string) => Promise<void> | void;
 };
-
-function ResolveClusterRow({
-  agent,
-  index,
-  total,
-  status,
-  threadComment,
-  diffComment,
-  isSelected,
-  isTaskActive,
-  canJump,
-  onSelect,
-  onJump,
-  onResolveThread,
-}: ResolveClusterRowProps) {
-  const hasUnread = agentHasUnread(agent, isSelected && isTaskActive);
-  const [pushing, setPushing] = useState(false);
-  const canPush = agent.sourceThreadId != null && (status === 'committed' || status === 'wontfix');
-  const onPush = async () => {
-    if (pushing || agent.sourceThreadId == null) {
-      return;
-    }
-    setPushing(true);
-    try {
-      await onResolveThread(agent.sourceThreadId);
-    } finally {
-      setPushing(false);
-    }
-  };
-  const snippet = threadComment ? (
-    <CommentSnippet
-      author={threadComment.author}
-      location={commentLocation(threadComment)}
-      body={threadComment.body}
-    />
-  ) : diffComment ? (
-    <CommentSnippet location={diffLocation(diffComment)} body={diffComment.body} />
-  ) : null;
-  return (
-    <div
-      role="button"
-      tabIndex={0}
-      onClick={onSelect}
-      onKeyDown={(e) => {
-        if (e.key === 'Enter' || e.key === ' ') {
-          e.preventDefault();
-          onSelect();
-        }
-      }}
-      aria-label={agent.name}
-      className={cn(
-        'relative flex w-full cursor-pointer flex-col gap-1 rounded border px-2 py-1.5 text-2xs font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-[var(--color-focus-ring)]',
-        isSelected
-          ? 'bg-elevated text-foreground border-border'
-          : 'text-foreground/70 hover:bg-muted/60',
-        status === 'running'
-          ? 'border-info/60'
-          : status === 'awaiting' || hasUnread
-            ? 'border-warning/70'
-            : 'border-transparent',
-      )}
-    >
-      <div className="flex w-full items-center gap-2">
-        <span className="tabular-nums text-muted-foreground/50">
-          {index + 1}/{total}
-        </span>
-        <span className="min-w-0 flex-1 truncate text-left">{agent.name}</span>
-        <ResolverStateBadge state={resolverBadgeState(status)} />
-        {canJump ? (
-          <button
-            type="button"
-            onClick={(e) => {
-              e.stopPropagation();
-              onJump();
-            }}
-            title="Go to comment"
-            aria-label="Go to comment"
-            className="shrink-0 rounded p-0.5 text-muted-foreground/60 transition-colors hover:bg-foreground/10 hover:text-foreground"
-          >
-            <MessageSquareReply size={11} aria-hidden />
-          </button>
-        ) : null}
-      </div>
-      {snippet ? <div className="pl-7">{snippet}</div> : null}
-      {canPush ? (
-        <div className="flex justify-end pl-7">
-          <button
-            type="button"
-            onClick={(e) => {
-              e.stopPropagation();
-              void onPush();
-            }}
-            disabled={pushing}
-            title="push the branch and resolve this comment now"
-            className={cn(
-              'inline-flex items-center gap-1 rounded-full border border-info/40 px-2 py-0.5 text-[10px] font-semibold text-info transition-colors hover:bg-info/10 disabled:cursor-not-allowed disabled:opacity-60',
-              pushing && 'animate-border-pulse',
-            )}
-          >
-            {pushing ? (
-              <StatusDot tone="info" size="sm" pulsing />
-            ) : (
-              <Upload size={9} aria-hidden />
-            )}
-            {pushing ? 'Pushing…' : 'Push & resolve this'}
-          </button>
-        </div>
-      ) : null}
-    </div>
-  );
-}
